pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faiss-cpu>=1.13.2",
    "numpy>=2.4.4",
    "openai>=2.36.0",
    "python-dotenv>=1.2.2",
]
